задание 5/grpc_app/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0

